requirements.txt
scikit-learn==1.5.2
SQLAlchemy==2.0.36
nltk==3.9.1
greek-stemmer-pos==1.1.2
spacy==3.8.2
tqdm==4.67.0
pandas==2.2.3
psycopg2==2.9.10
Flask-SQLAlchemy==3.1.1
flask==3.1.0
matplotlib==3.9.2
faiss-cpu==1.9.0.post1
hdbscan==0.8.40
seaborn==0.13.2
umap==0.1.1
tensorflow==2.18.0
umap-learn==0.5.7
transformers==4.46.3
python-dotenv==1.0.1
numpy~=2.0.2
scipy~=1.14.1